add test for simple measurement widget
