+++ package/tests/test_qt_part/test_io.py
@@ -4,9 +4,6 @@ from pathlib import Path
 
 import pytest
 
-from PartSeg.segmentation_analysis.partseg_settings import PartSettings
-from PartSeg.segmentation_mask.main_window import ChosenComponents
-from PartSeg.segmentation_mask.stack_settings import StackSettings
 from PartSegCore.analysis import ProjectTuple
 from PartSegCore.analysis.load_functions import LoadImageMask, LoadMaskSegmentation, LoadProject, LoadStackImage
 from PartSegCore.analysis.save_functions import save_dict as analysis_save_dict
@@ -18,30 +15,6 @@ from PartSegCore.mask.io_functions import (
     save_parameters_dict,
     save_segmentation_dict,
 )
-
-
-@pytest.fixture
-def part_settings(image, tmp_path):
-    settings = PartSettings(tmp_path)
-    settings.image = image
-    return settings
-
-
-@pytest.fixture
-def stack_settings(qtbot, tmp_path):
-    settings = StackSettings(tmp_path)
-    chose = ChosenComponents()
-    qtbot.addWidget(chose)
-    settings.chosen_components_widget = chose
-    return settings
-
-
-@pytest.fixture
-def part_settings_with_project(image, analysis_segmentation2, tmp_path):
-    settings = PartSettings(tmp_path)
-    settings.image = image
-    settings.set_project_info(analysis_segmentation2)
-    return settings
 
 
 class TestPartSettingsIO:
